pyproject.toml
[project]
name = "ai-resume-critiquer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "openai>=2.7.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
